Move test helpers to separate file to declutter the test files
