+++ src/test/scala/calc/RunTest.scala
@@ -1,15 +1,7 @@
 package calc
 
 import org.junit.Test
-import org.junit.Assert._
-
-import org.scalajs.core.ir
-import ir.{Trees => irt, Types => irtpe}
-import ir.Definitions._
-
-import org.scalajs.core.tools.logging._
-
-import org.scalajs.jsenv.JSConsole
+import TestHelpers._
 
 /** End-to-end tests, with parsing, compiling and running.
  *
@@ -17,24 +9,9 @@ import org.scalajs.jsenv.JSConsole
  *  compiling and running some piece of code produces the expected final
  *  result.
  */
-class RunTest {
+class RunSimpleValue {
 
-  private def assertRun(expected: Double, code: String): Unit = {
-    val tree = Parser.parse(code).get.value
-    val classDef = Compiler.compileMainClass(tree)
-    val linked = Linker.link(classDef, NullLogger)
-
-    val lines = new java.io.StringWriter
-    val console = new JSConsole {
-      def log(msg: Any): Unit = lines.append(msg.toString() + "\n")
-    }
-
-    Runner.run(linked, NullLogger, console)
-
-    assertEquals(expected.toString(), lines.toString().trim)
-  }
-
-  @Test def runLiteral(): Unit = {
+  @Test def literal(): Unit = {
     assertRun(54.3, "54.3")
   }
