+++ src/test/scala/calc/TestHelpers.scala
@@ -1,8 +1,48 @@
 package calc
 
-/**
-  * Created by evan on 25/3/16.
-  */
-class TestHelpers {
+import org.junit.Assert._
+import org.scalajs.core.ir
+import org.scalajs.core.ir.{Trees => irt, Types => irtpe}
+import org.scalajs.core.tools.logging.NullLogger
+import org.scalajs.jsenv.JSConsole
 
+object TestHelpers {
+
+  implicit val DummyPos = ir.Position.NoPosition
+
+  def assertRun(expected: Double, code: String): Unit = {
+    val tree = Parser.parse(code).get.value
+    val classDef = Compiler.compileMainClass(tree)
+    val linked = Linker.link(classDef, NullLogger)
+
+    val lines = new java.io.StringWriter
+    val console = new JSConsole {
+      def log(msg: Any): Unit = lines.append(msg.toString() + "\n")
+    }
+
+    Runner.run(linked, NullLogger, console)
+
+    assertEquals(expected.toString(), lines.toString().trim)
+  }
+
+  def assertCompile(expected: irt.Tree, sourceTree: Tree): Unit = {
+    /* IR Trees do not have a meaningful equals() method, so we test equality
+     * through hashes.
+     */
+
+    def hashOf(body: irt.Tree): irt.TreeHash = {
+      // Can only hash entire methods
+      val methodDef = irt.MethodDef(static = false, irt.Ident("main__D"),
+        Nil, irtpe.DoubleType, body)(
+        irt.OptimizerHints.empty, None)
+      ir.Hashers.hashMethodDef(methodDef).hash.get
+    }
+
+    val expectedHash = hashOf(expected)
+    val actual = Compiler.compileExpr(sourceTree)
+    val actualHash = hashOf(actual)
+
+    assertTrue(s"Expected $expected but got $actual",
+      ir.Hashers.hashesEqual(actualHash, expectedHash, considerPos = true))
+  }
 }
